backend/pyproject.toml
[project]
name = "interactive-city-map"
version = "0.1.0"
description = "Multi-layer interactive city map with historical overlays and building-level interactivity"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy[asyncio]==2.0.35",
    "geoalchemy2==0.15.2",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.3",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-multipart==0.0.12",
    "httpx==0.27.2",
    "boto3==1.35.36",
    "shapely==2.0.6",
    "fiona==1.10.1",
    "pyproj==3.7.0",
    "geojson==3.1.0",
    "structlog==24.4.0",
    "pyyaml==6.0.2",
    "python-dotenv==1.0.1",
    # Interior maps dependencies
    "opencv-python-headless==4.10.0.84",
    "pytesseract==0.3.13",
    "beautifulsoup4==4.12.3",
    "playwright==1.49.1",
    "scikit-learn==1.5.2",
    "Pillow==10.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "httpx==0.27.2",
    "black==24.8.0",
    "ruff==0.6.9",
    "mypy==1.11.2",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
plugins = ["sqlalchemy.ext.mypy.plugin"]
